% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_ScRNA.R
\name{Cluster_ScRNA}
\alias{Cluster_ScRNA}
\title{Title Calculate Cluster information}
\usage{
Cluster_ScRNA(x = scRNA, res = c(seq(0.1, 1.6, 0.2)), npcs = 30, dims = 1:30)
}
\arguments{
\item{x}{scRNA}

\item{res}{resolution that you want to cluster}

\item{npcs}{the param of runPCA}

\item{dims}{which numbers of npcs you want to use}
}
\value{
ScRNA that be cluster
}
\description{
Title Calculate Cluster information
}
