% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareVirusCB.R
\name{CompareVirusCB}
\alias{CompareVirusCB}
\title{Title Used to check the proportion of viruses in SCRNA data}
\usage{
CompareVirusCB(vct = vclist, sct = sclist)
}
\arguments{
\item{vct}{vclist input}

\item{sct}{sclist input}
}
\description{
{in: vclist} {out: print(table)}
}
\author{
Jesse
}
