% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareVirusCB.R
\name{CompareVirusCB.vclist}
\alias{CompareVirusCB.vclist}
\title{Title vclist compare}
\usage{
\method{CompareVirusCB}{vclist}(vct, sct = sclist)
}
\arguments{
\item{vct}{vclist input}

\item{sct}{sclist input}
}
\value{
table
}
\description{
Title vclist compare
}
