% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareVirusCB.R
\name{CompareVirusCB.vcsingle}
\alias{CompareVirusCB.vcsingle}
\title{Title single vclist compare}
\usage{
\method{CompareVirusCB}{vcsingle}(vct, sct = sclist)
}
\arguments{
\item{vct}{single vclist input}

\item{sct}{sclist input}
}
\value{
table
}
\description{
Title single vclist compare
}
