% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViruSeurat.R
\name{.onAttach}
\alias{.onAttach}
\title{Title loaded the package what todo}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{libname}

\item{pkgname}{pkgname}
}
\value{

}
\description{
Title loaded the package what todo
}
\examples{

}
