% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_testdata.R
\name{Input_Testdata}
\alias{Input_Testdata}
\title{It will generat a Test Data}
\usage{
Input_Testdata(pos = 1)
}
\arguments{
\item{pos}{enviroment position}
}
\description{
It will generat a Test Data
}
