% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_data.R
\name{Input_Vclist}
\alias{Input_Vclist}
\title{Title Use to input Virus counts data to Vclist,{in: $sample viralname } {out: $(sample)sum}}
\usage{
Input_Vclist(x = viruscount, inputfiledir = "data/", viralname = "NC_003977")
}
\arguments{
\item{x}{give dataset that contail viruscount}

\item{inputfiledir}{input Virus counts data dir}

\item{viralname}{give a viralname to filter and input to vclist}
}
\value{
vclist
}
\description{
defaultly,it will automatic detect Data in data/ directory.
The Virus-data type is like "CRR703.Exp.tsv" generate from Viral-track.
viralname is which virus you want to count for.
vclist<-Input_Vclist()  #it will automatic detect Data in data/ directory.
vclist<-Input_Vclist(x=viruscount,viralname="NC_003977") #you can use your own viruscount dataset and specific viralname.
}
