% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_IntegrateSclist.R
\name{Normalize_IntegrateSclist}
\alias{Normalize_IntegrateSclist}
\title{Title Normalize and Integrate Svclist, generate a scRNA : an integrated Seurat object}
\usage{
Normalize_IntegrateSclist(x = svclist, fmethod = "vst", features = 2000)
}
\arguments{
\item{x}{svclist which has virus counts, it is a sclist object}

\item{fmethod}{method use to calculate VariableFeatures}

\item{features}{use to find High variable genes}
}
\value{
scRNA
}
\description{
Title Normalize and Integrate Svclist, generate a scRNA : an integrated Seurat object
}
