% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Cluster.R
\name{Plot_Cluster}
\alias{Plot_Cluster}
\title{Title Plot Cluster by orig.ident/sample/patient in your specific resolution}
\usage{
Plot_Cluster(x = ccRNA, res = 0.5)
}
\arguments{
\item{x}{input ccRNA}

\item{res}{resolution you want to plot}
}
\value{
Cluster Plot
}
\description{
Title Plot Cluster by orig.ident/sample/patient in your specific resolution
}
