% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Scatter.R
\name{Plot_Scatter}
\alias{Plot_Scatter}
\title{Title Plot_scatter in integrated Seurat object, find corresponding between virus and cells}
\usage{
Plot_Scatter(x = scRNA)
}
\arguments{
\item{x}{scRNA that is integrated Seurat object}
}
\value{
Plot object
}
\description{
Title Plot_scatter in integrated Seurat object, find corresponding between virus and cells
}
