% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_VirusCount.R
\name{Plot_VirusCount}
\alias{Plot_VirusCount}
\title{Title Read number distribution histogram for drawing the imported viral data}
\usage{
Plot_VirusCount(x = vclist, meta = metadata)
}
\arguments{
\item{x}{vclist you want to count from}

\item{meta}{metadata}
}
\value{
wrap-plot-list
}
\description{
defaultly,it will automatic detect Data (vclist/metadata) in global env.
This function is used to plot virusCount in your dataset, you can visualize your VirusCounts distribution .
p1<-Plot_VirusCount()  # it will automatic detect Data (vclist/metadata) in global env.
Plot_VirusCount(x=vclist,meta=metadata) # set your own parameter .
}
