% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViruSeurat.R
\name{Print_attach}
\alias{Print_attach}
\title{Print_attach}
\usage{
Print_attach()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Print_attach when attach
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
