% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViruSeurat.R
\name{Print_dir}
\alias{Print_dir}
\title{Print_dir}
\usage{
Print_dir()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Print_dir when attach
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
