% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViruSeurat.R
\name{Print_intro}
\alias{Print_intro}
\title{Print_intro}
\usage{
Print_intro()
}
\value{
intro
}
\description{
print_intro when attach
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
