% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViruSeurat.R
\name{Print_logo}
\alias{Print_logo}
\title{Print_logo}
\usage{
Print_logo()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Print_logo when attach
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{

}
