% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recluster_CcRNA.R
\name{Recluster_CcRNA}
\alias{Recluster_CcRNA}
\title{Title ReCalculate Cluster information}
\usage{
Recluster_CcRNA(x = ccRNA, res = c(seq(0.1, 1.6, 0.2)))
}
\arguments{
\item{x}{ccRNA}

\item{res}{new resolution}
}
\value{
new Recluster ccRNA
}
\description{
Title ReCalculate Cluster information
}
