% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VirusinSclist.R
\name{VirusinSclist}
\alias{VirusinSclist}
\title{Title Input Virus Counts to Seurat-list}
\usage{
VirusinSclist(x = vclist, sct = sclist, meta = metadata)
}
\arguments{
\item{x}{vclist}

\item{sct}{sclist}

\item{meta}{metadata}
}
\value{
svclist
}
\description{
svclist<-ViralinSeurat() # This step is put Virus counts in Sclist, generat a Svclist.
}
